requirements.txt
meshtastic>=2.3.0
PyYAML>=6.0
pypubsub>=4.0.3
